+++ master_protoactor/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+__version__ = "0.0.1"
