+++ master_protoactor/pid.py
@@ -1,0 +1,5 @@
+class PID(object):
+    def __init__(self, address, id, ref=None):
+        self.address = address
+        self.id = id
+        self.ref = ref
