+++ tests/test_deps.py
@@ -4,10 +4,10 @@
 
 def test_circular_dependencies():
     """Verify that there are no circular dependencies"""
-    from protoactor.utils import *
-    from protoactor.messages import *
-    from protoactor.mailbox import *
-    from protoactor.pid import *
-    from protoactor.process import *
-    from protoactor.processregistry import *
-    from protoactor.props import *
+    from master_protoactor.utils import *
+    from master_protoactor.messages import *
+    from master_protoactor.mailbox import *
+    from master_protoactor.pid import *
+    from master_protoactor.process import *
+    from master_protoactor.processregistry import *
+    from master_protoactor.props import *
